go-1436: create dockerfile
Create a Dockerfile to allow OpenPedigree to be run within a Docker
container.  Update README to provide brief usage instructions.  Add
additional line to package.json for docker-triggered `start` command.

I tried using the Alpine linux version of the NodeJS image, but this
produced errors, so I've stuck with the stock version.